pyproject.toml
[tool.poetry]
name = "sitesavant"
version = "1.0.0"
description = ""
authors = ["Piotr Ząbek <p.zabek01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"
pinecone-client = "^2.2.4"
h5py = "^3.10.0"
faiss-cpu = "^1.7.4"
langchain = "^0.1.3"
langchain-openai = "^0.0.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^24.1.1"
pytest-mock = "^3.12.0"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sitesavant = "app.main:main"

[tool.black]
line-length = 99
